+++ components/SettingsProvider.tsx
@@ -3,23 +3,12 @@ import React, { useEffect } from "react";
 import { cookieSettings, Settings } from "@/shared/settings";
 import { usePageContext } from "vike-react/usePageContext";
 import Cookies from "js-cookie";
+import { SettingsContext, SettingsContextContent } from "@/hooks/useSettings";
 
 const cookies = Cookies.withAttributes({
   sameSite: cookieSettings.sameSite,
   secure: cookieSettings.secure,
   expires: cookieSettings.maxAgeDays,
-});
-
-type SettingsContextContent = {
-  initialized: boolean;
-  settings: Settings;
-  setSettings: React.Dispatch<React.SetStateAction<Settings>>;
-};
-
-const SettingsContext = React.createContext<SettingsContextContent>({
-  initialized: false,
-  settings: Settings.default,
-  setSettings() {},
 });
 
 export type SettingsProviderProps = {
@@ -48,15 +37,4 @@ export function SettingsProvider(props: SettingsProviderProps) {
       {props.children}
     </SettingsContext.Provider>
   );
-}
-
-export function useSettings() {
-  const { initialized, settings, setSettings } =
-    React.useContext(SettingsContext);
-
-  if (!initialized) {
-    throw new Error("Attempting to use settings outside <SettingsProvider>.");
-  }
-
-  return [settings, setSettings] as const;
 }
